requirements.txt
pytest-selenium
pytest-rerunfailures
pytest-xdist
allure-pytest
requests
pyyaml
pytest
pytest_bdd